t7900: make macOS-specific test work on Windows
Although `git maintenance start` and `git maintenance stop` necessarily
invoke platform-specific scheduling utilities, their related tests have
been carefully crafted -- with one minor exception -- to work correctly
on any platform, thus improving overall coverage. The exception is that
the macOS-specific test fails on Windows due to non-portable use of
`$(id -u)` and comparison involving the value of $HOME.

In particular, on Windows, the value of getuid() called by the C code is
not guaranteed to be the same as `$(id -u)` invoked by the test. This is
because `git.exe` is a native Windows program, whereas the utility
programs run by the test script mostly utilize the MSYS2 runtime, which
emulates a POSIX-like environment. Since the purpose of the test is to
check that the input to the hook is well-formed, the actual user ID is
immaterial, thus we can work around the problem by making the the test
UID-agnostic.

As for comparison of $HOME, it suffers from the typical shortcoming on
Windows in which the same path may be represented two different ways
depending upon its source (i.e. as a Windows path
`C:/git-sdk-64/usr/src/git/foo` versus as a Unix path
`/usr/src/git/foo`).

Fix both problems and drop the !MINGW prerequisite from the
macOS-specific test, thus allowing the test to run on Windows, as well.

Helped-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Eric Sunshine <sunshine@sunshineco.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>